refactor routes into svc package
