+++ src/main/scala/com/brian/funkyserver/FunkyserverServer.scala
@@ -1,17 +1,20 @@
 package com.brian.funkyserver
 
 import cats.effect.{ConcurrentEffect, ContextShift, Timer}
-import cats.implicits._
 import fs2.Stream
 import org.http4s.client.blaze.BlazeClientBuilder
-import org.http4s.implicits._
 import org.http4s.server.blaze.BlazeServerBuilder
 import org.http4s.server.middleware.Logger
 import scala.concurrent.ExecutionContext.global
 
+import cats.implicits._
+import org.http4s.implicits._
+
+import svc.{FunkyserverService, HelloWorld, Jokes}
+
 object FunkyserverServer {
 
-  def stream[F[_]: ConcurrentEffect](implicit T: Timer[F], C: ContextShift[F]): Stream[F, Nothing] = {
+  def stream[F[_] : ConcurrentEffect](implicit T: Timer[F], C: ContextShift[F]): Stream[F, Nothing] = {
     for {
       client <- BlazeClientBuilder[F](global).stream
       helloWorldAlg = HelloWorld.impl[F]
@@ -22,9 +25,9 @@ object FunkyserverServer {
       // want to extract a segments not checked
       // in the underlying routes.
       httpApp = (
-        FunkyserverRoutes.helloWorldRoutes[F](helloWorldAlg) <+>
-        FunkyserverRoutes.jokeRoutes[F](jokeAlg)
-      ).orNotFound
+        FunkyserverService.helloWorldRoutes[F](helloWorldAlg) <+>
+          FunkyserverService.jokeRoutes[F](jokeAlg)
+        ).orNotFound
 
       // With Middlewares in place
       finalHttpApp = Logger.httpApp(true, true)(httpApp)
